Add detection of thread creation rather than assuming it is a fork'd process
Unfortunately a clone() system call will end up with the ppid
referencing the parent of the process doing the clone and not the pid of
process actually doing the cloning.  So, instead, parse /proc/$pid/status
for the ppid and tid and deduce from that if the new process is a real
child from a fork, or a thread from a clone.

The benefit is now that forkstat identifies parent processes correctly
and we can also print out info stating of the new process is a thread
or a child.

Thanks to sturmflut@lieberbiber.de for spotting this issue and helping
to diagnose what was going wrong.

Signed-off-by: Colin Ian King <colin.king@canonical.com>