Re-work unknown cmdline information
Currently the cmdline information on unknown process names is
rather poorly handled.   Don't allocate or free an unknown
cmdline, unstead use a static unknown string to save heap
activity and string duplication.  Also opportunistically
try to keep the cmdline updated especially if it is unknown.
Don't update a known cmdline if the new information will
make it unknown.  Finally, because pid 0 (swapper) does not
have a cmdline, replace it with "[swapper]" to at least
indicate it is a kernel 0 process.

Signed-off-by: Colin Ian King <colin.king@canonical.com>